Update README.md
Value Returning Square Function

Write a program that prompts the user for a number and prints that number's square to the console.

Your program should define a function called square that has one floating-point variable named num as a parameter. When the function executes, it should return the square of the num parameter.

The program in the main function should prompt the user for a number, then call the square function with that number as its argument and assign its return value to a variable that will hold the result. Finally, the result should be displayed to the console as shown in the sample run.

Sample run:

Enter a number to square: 4.2
The square is 17.64